refactor(core): extract shared presentation logic into core.ts
- Create generator/core.ts with generatePresentationCore function
- Add ChartRenderingContext interface for dependency injection
- Replace console.warn with optional onChartRenderError callback
- Reduce index.ts from 186 to 39 lines
- Reduce browser.ts from 207 to 51 lines
- Eliminate ~150 lines of code duplication
